% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasma_contamination.R
\name{plasma_contamination}
\alias{plasma_contamination}
\title{Ipsum...}
\usage{
plasma_contamination(df, n_perm = 1000)
}
\arguments{
\item{df}{ipsum...}
}
\value{
ipsum...
}
\description{
Ipsum...
}
